requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
pydantic==2.5.0
python-multipart==0.0.6
boto3>=1.34.72
langchain>=0.1.0
langchain-aws>=0.1.0
langchain-community>=0.0.10
python-dateutil==2.8.2
icalendar==5.0.11
google-api-python-client==2.110.0
google-auth-httplib2==0.2.0
google-auth-oauthlib==1.1.0
# Voice processing dependencies (optional)
# speechrecognition==3.10.0
# pydub==0.25.1
# pyaudio==0.2.11
# openai-whisper==20231117
numpy==1.24.3
pandas==2.1.4
python-dotenv==1.0.0
cors==1.0.1
fastapi-cors==0.0.6